fix : Manager인 경우에는 findCustomer할 때 error 처리 (접근 못하게 막음)

+++ src/main/java/com/blossom/lineup/Waiting/service/WaitingServiceImpl.java
@@ -55,6 +55,10 @@ public class WaitingServiceImpl implements WaitingService{
 
     private Customer findCustomer(){
         CustomUserDetails currentUserInfo = securityUtils.getCurrentUserInfo();
+
+        if(currentUserInfo.getRole().equals("MANAGER")){ // 매니저는 유저"용" 대기 접근 할 수 없음!!
+            throw new BusinessException(Code.MANAGER_ACCESS_DENIED);
+        }
         return customerRepository.findById(currentUserInfo.getId())
                 .orElseThrow(()-> new BusinessException(Code.CUSTOMER_NOT_FOUND));
     }
@@ -89,6 +93,12 @@ public class WaitingServiceImpl implements WaitingService{
         log.debug(waitingId+" 대기를 삭제했습니다.");
 
         Waiting w = findWaiting(waitingId);
+
+        Customer customer = findCustomer();
+        if(w.getCustomer().getId().equals(customer.getId())){ // 지우기 전 권한 검사.
+            throw new BusinessException(Code.WAITING_NOT_MATCH_USER);
+        }
+
         waitingRepository.delete(w);
     }
 
@@ -136,7 +146,7 @@ public class WaitingServiceImpl implements WaitingService{
         Customer customer = findCustomer();
         log.info("qr-code 요청 유저 {} : waiting 유저 {}", customer.getId(),waiting.getCustomer().getId());
 
-        if(waiting.getCustomer().getId() != customer.getId()){
+        if(waiting.getCustomer().getId().equals(customer.getId())){
             throw new BusinessException(Code.WAITING_NOT_MATCH_USER);
         }
 
@@ -151,9 +161,6 @@ public class WaitingServiceImpl implements WaitingService{
     /**
      * 대기 현황 조회
      * (WAITING 상태일 때)
-     * @param waiting
-     * @param organizationId
-     * @return
      */
     private WaitingResponse myCurrentWaiting(Waiting waiting, Long organizationId) {
 
@@ -175,22 +182,22 @@ public class WaitingServiceImpl implements WaitingService{
         List<Waiting> usingTables = waitingRepository.findUsingTables(o);
 
         // 정렬된 각각의 테이블의 남은 이용 시간(분)
-        List<Integer> remainTableTimes = new java.util.ArrayList<>(usingTables.stream()
-            .flatMap(w ->{
-                // 대기시간이 null인 테이블이 존재하면, exception
-                if(w.getEntranceTime()==null){
-                    throw new BusinessException(Code.ENTRANCE_TIME_IS_NULL);
-                } else {
-                    LocalDateTime entranceTime = w.getEntranceTime();
-                    LocalDateTime currentTime = LocalDateTime.now();
-                    Duration duration = Duration.between(entranceTime,currentTime);
-                    int minutes = (int) duration.toMinutes(); // 분단위로 변환
-                    int remainMinutes = Math.max(o.getTableTimeLimit() - minutes, 0);
+        List<Integer> remainTableTimes = usingTables.stream()
+                .flatMap(w -> {
+                    // 대기시간이 null인 테이블이 존재하면, exception
+                    if (w.getEntranceTime() == null) {
+                        throw new BusinessException(Code.ENTRANCE_TIME_IS_NULL);
+                    } else {
+                        LocalDateTime entranceTime = w.getEntranceTime();
+                        LocalDateTime currentTime = LocalDateTime.now();
+                        Duration duration = Duration.between(entranceTime, currentTime);
+                        int minutes = (int) duration.toMinutes(); // 분단위로 변환
+                        int remainMinutes = Math.max(o.getTableTimeLimit() - minutes, 0);
 
-                    return java.util.stream.IntStream.range(0, w.getTableCount())
-                            .mapToObj(i -> remainMinutes); // 각 tableCnt 개수만큼 지속시간을 반복
-                }
-            }).collect(Collectors.toList()));
+                        return java.util.stream.IntStream.range(0, w.getTableCount())
+                                .mapToObj(i -> remainMinutes); // 각 tableCnt 개수만큼 지속시간을 반복
+                    }
+                }).collect(Collectors.toList());
 
         // 주점에서 다루는 테이블보다 테이블 이용 개수가 적으면, List에 0분 남은 개수만큼 추가.
         while(remainTableTimes.size() < o.getTableCount()){
@@ -210,8 +217,6 @@ public class WaitingServiceImpl implements WaitingService{
     /**
      * 입장중 상태 조회
      * (PENDING)
-     * @param waiting
-     * @return
      */
     private PendingResponse getPendingStatus(Waiting waiting) {
         String waitingStatus = EntranceStatus.PENDING.getEntranceStatus();
